srv-lib: Add explicit pollint cast for sign conv
Silences -Wsign-conversion

+++ src/prg/c/src/srv-lib/libswocserver/optionproc.c
@@ -8,7 +8,7 @@
  * Released under the GPLv3 only.\n
  * SPDX-License-Identifier: GPL-3.0
  *
- * @version _v1.1.10 ==== 18/05/2019_
+ * @version _v1.1.11 ==== 02/06/2019_
  */
 
 /* **********************************************************************
@@ -56,6 +56,9 @@
  *				bool.					*
  * 07/09/2018	MG	1.1.9	Use new mgemessage struct initialiser.	*
  * 18/05/2019	MG	1.1.10	Merge sub-projects into one.		*
+ * 02/06/2019	MG	1.1.11	Add explicit cast to pollint for sign	*
+ *				conversion. It has already been		*
+ *				validated > 0.				*
  *									*
  ************************************************************************
  */
@@ -369,7 +372,7 @@ int sws_server_wait(void)
 	if (prg_err)
 		return prg_err;
 	do {
-		sleep(pollint);
+		sleep((unsigned int)pollint);
 		prg_err = exch_msg(out_msg, om_length, msg);
 		if (prg_err)
 			return prg_err;
@@ -397,7 +400,7 @@ int sws_server_wait(void)
 		clear_msg(msg, ';', ',');
 	} while (empty);
 	/* Allow one more sleep as a buffer, maybe client is shutting down. */
-	sleep(pollint);
+	sleep((unsigned int)pollint);
 	syslog((int) (LOG_USER | LOG_NOTICE), "No outstanding client locks on "
 		"this server.");
 	return 0;
